healix-app/src/python_service/requirements.txt
flask
flask-cors
deepface
tf-keras
psycopg2-binary
requests
gunicorn
